Back/Electrovia/bin/Debug/net8.0/python/requirements.txt
numpy
faiss-cpu
Pillow
tqdm
torch
torchvision
